Move the noise utilities
