+++ src/lightcurvelynx/example_runs/simulate_snia.py
@@ -4,7 +4,6 @@ from pathlib import Path
 
 import numpy as np
 from astropy import units as u
-from lightcurvelynx.astro_utils.noise_model import apply_noise
 from lightcurvelynx.astro_utils.passbands import PassbandGroup
 from lightcurvelynx.astro_utils.snia_utils import (
     DistModFromRedshift,
@@ -17,6 +16,7 @@ from lightcurvelynx.math_nodes.np_random import NumpyRandomFunc
 from lightcurvelynx.math_nodes.scipy_random import SamplePDF
 from lightcurvelynx.models.sncosmo_models import SncosmoWrapperModel
 from lightcurvelynx.models.snia_host import SNIaHost
+from lightcurvelynx.noise_models.noise_utils import apply_noise
 from scipy.interpolate import interp1d
 
 logger = logging.getLogger(__name__)
