+++ src/lightcurvelynx/noise_models/noise_utils.py
