+++ src/lightcurvelynx/obstable/skymapper_obstable.py
@@ -8,8 +8,8 @@ from citation_compass import CiteClass
 from lightcurvelynx.astro_utils.coordinate_utils import build_moc_from_coords
 from lightcurvelynx.astro_utils.detector_footprint import DetectorFootprint
 from lightcurvelynx.astro_utils.mag_flux import mag2flux
-from lightcurvelynx.astro_utils.noise_model import poisson_bandflux_std
 from lightcurvelynx.consts import GAUSS_EFF_AREA2FWHM_SQ
+from lightcurvelynx.noise_models.noise_utils import poisson_bandflux_std
 from lightcurvelynx.obstable.obs_table import ObsTable
 
 SKYMAPPER_PIXEL_SCALE = 0.497
