+++ src/lightcurvelynx/simulate.py
@@ -9,8 +9,8 @@ import pandas as pd
 from nested_pandas import NestedFrame
 from tqdm import tqdm
 
-from lightcurvelynx.astro_utils.noise_model import apply_noise
 from lightcurvelynx.astro_utils.spectrograph import Spectrograph
+from lightcurvelynx.noise_models.noise_utils import apply_noise
 from lightcurvelynx.utils.post_process_results import concat_results
 
 logger = logging.getLogger(__name__)
@@ -446,10 +446,24 @@ def _simulate_lightcurves_batch(simulation_info):
                     state,
                     rng_info=rng,
                 )
-                bandfluxes_error = obstable[survey_idx].bandflux_error_point_source(
-                    bandfluxes_perfect, obs_index
-                )
-                bandfluxes = apply_noise(bandfluxes_perfect, bandfluxes_error, rng=rng)
+
+                # During the code transition we support both the object
+                # based noise models and the ObsTable implementation of noise.
+                # TODO: Remove the else case when we have moved everything
+                # to the object-based noise models.
+                if obstable[survey_idx].noise_model is not None:
+                    noise_model = obstable[survey_idx].noise_model
+                    bandfluxes, bandfluxes_error = noise_model.apply_noise(
+                        bandfluxes_perfect,
+                        obs_index,
+                        rng=rng,
+                    )
+                else:
+                    bandfluxes_error = obstable[survey_idx].bandflux_error_point_source(
+                        bandfluxes_perfect,
+                        obs_index,
+                    )
+                    bandfluxes = apply_noise(bandfluxes_perfect, bandfluxes_error, rng=rng)
 
                 # Apply saturation thresholds from the ObsTable if they are requested.
                 if simulation_info.apply_saturation:
